# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "API Documentation",
      "description": "API endpoints for User, Invoice, Subscription, and Payments"
    },
    "basePath": "/api",
    "schemes": [
      "http",
      "https"
    ],
    "paths": {
      "/users": {
        "get": {
          "tags": [
            "User"
          ],
          "summary": "Get all users",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "User"
          ],
          "summary": "Create a new user",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "user",
              "description": "User object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "/users/{id}": {
        "get": {
          "tags": [
            "User"
          ],
          "summary": "Get a user by ID",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "User ID",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        },
        "put": {
          "tags": [
            "User"
          ],
          "summary": "Update a user by ID",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "User ID",
              "required": true,
              "type": "string",
              "format": "uuid"
            },
            {
              "in": "body",
              "name": "user",
              "description": "Updated user object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          }
        },
        "delete": {
          "tags": [
            "User"
          ],
          "summary": "Delete a user by ID",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "User ID",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            }
          }
        }
      },
      "/invoices": {
        "get": {
          "tags": [
            "Invoice"
          ],
          "summary": "Get all invoices",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Invoice"
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Invoice"
          ],
          "summary": "Create a new invoice",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "invoice",
              "description": "Invoice object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Invoice"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Invoice"
              }
            }
          }
        }
      },
      "/invoices/{id}": {
        "get": {
          "tags": [
            "Invoice"
          ],
          "summary": "Get an invoice by ID",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Invoice ID",
              "required": true,
              "type": "string",
              "format": "uuid"

            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Invoice"
              }
            }
          }
        },
        "put": {
          "tags": [
            "Invoice"
          ],
          "summary": "Update an invoice by ID",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Invoice ID",
              "required": true,
              "type": "string",
              "format": "uuid"
            },
            {
              "in": "body",
              "name": "invoice",
              "description": "Updated invoice object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Invoice"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Invoice"
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Invoice"
          ],
          "summary": "Delete an invoice by ID",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Invoice ID",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            }
          }
        }
      },
      "/subscriptions": {
        "get": {
          "tags": [
            "Subscription"
          ],
          "summary": "Get all subscriptions",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Subscription"
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Subscription"
          ],
          "summary": "Create a new subscription",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "subscription",
              "description": "Subscription object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Subscription"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Subscription"
              }
            }
          }
        }
      },
      "/subscriptions/{id}": {
        "get": {
          "tags": [
            "Subscription"
          ],
          "summary": "Get a subscription by ID",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Subscription ID",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Subscription"
              }
            }
          }
        },
        "put": {
          "tags": [
            "Subscription"
          ],
          "summary": "Update a subscription by ID",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Subscription ID",
              "required": true,
              "type": "string",
              "format": "uuid"
            },
            {
              "in": "body",
              "name": "subscription",
              "description": "Updated subscription object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Subscription"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Subscription"
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Subscription"
          ],
          "summary": "Delete a subscription by ID",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Subscription ID",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            }
          }
        }
      },
      "/payments": {
        "get": {
          "tags": [
            "Payment"
          ],
          "summary": "Get all payments",
          "produces": [
            "application/json"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Payment"
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Payment"
          ],
          "summary": "Create a new payment",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "payment",
              "description": "Payment object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Payment"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Payment"
              }
            }
          }
        }
      },
      "/payments/{id}": {
        "get": {
          "tags": [
            "Payment"
          ],
          "summary": "Get a payment by ID",
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Payment ID",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Payment"
              }
            }
          }
        },
        "put": {
          "tags": [
            "Payment"
          ],
          "summary": "Update a payment by ID",
          "consumes": [
            "application/json"
          ],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Payment ID",
              "required": true,
              "type": "string",
              "format": "uuid"
            },
            {
              "in": "body",
              "name": "payment",
              "description": "Updated payment object",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Payment"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Payment"
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Payment"
          ],
          "summary": "Delete a payment by ID",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "description": "Payment ID",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            }
          }
        }
      },
      "/invites": {
        "get": {
          "summary": "Get all invites",
          "tags": [
            "Invite"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Invite"
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new invite",
          "tags": [
            "Invite"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Invite object that needs to be created",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Invite"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created",
              "schema": {
                "$ref": "#/definitions/Invite"
              }
            }
          }
        }
      },
      "/invites/{id}": {
        "get": {
          "summary": "Get an invite by ID",
          "tags": [
            "Invite"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the invite",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Invite"
              }
            }
          }
        },
        "put": {
          "summary": "Update an invite",
          "tags": [
            "Invite"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the invite",
              "required": true,
              "type": "string",
              "format": "uuid"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Invite object that needs to be updated",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Invite"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Invite"
              }
            }
          }
        },
        "delete": {
          "summary": "Delete an invite",
          "tags": [
            "Invite"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the invite",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            }
          }
        }
      },
      "/photos": {
        "get": {
          "summary": "Get all photos",
          "tags": [
            "Photo"
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/Photo"
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new photo",
          "tags": [
            "Photo"
          ],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "description": "Photo object that needs to be created",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Photo"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created",
              "schema": {
                "$ref": "#/definitions/Photo"
              }
            }
          }
        }
      },
      "/photos/{id}": {
        "get": {
          "summary": "Get a photo by ID",
          "tags": [
            "Photo"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the photo",
              "required": true,
              "type": "string",
              "format": "uuid"            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Photo"
              }
            }
          }
        },
        "put": {
          "summary": "Update a photo",
          "tags": [
            "Photo"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the photo",
              "required": true,
              "type": "string",
              "format": "uuid"
            },
            {
              "in": "body",
              "name": "body",
              "description": "Photo object that needs to be updated",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Photo"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/Photo"
              }
            }
          }
        },
        "delete": {
          "summary": "Delete a photo",
          "tags": [
            "Photo"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "ID of the photo",
              "required": true,
              "type": "string",
              "format": "uuid"
            }
          ],
          "responses": {
            "204": {
              "description": "No Content"
            }
          }
        }
      }
    },
    "definitions": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "google_id": {
            "type": "string"
          },
          "apple_id": {
            "type": "string"
          },
          "microsoft_id": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum":[
              "CUSTOMER",
              "ADMIN",
              "SUPERADMIN",
              "EXECUTIVEBOARDMEMBER"
            ]
          },
          "status": {
            "type": "string",
            "enum":[
              "ACTIVE",
              "INVITED",
              "DISABLED",
              "INACTIVE"
            ]
          }
        }
      },
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "charges": {
            "type": "number",
            "format": "double"
          },
          "discounts": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string"
          },
          "due_date": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "SENT",
              "VIEWED",
              "PAID"
            ]
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Subscription": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "plan_name": {
            "type": "string",
            "enum":[
              "STANDARD",
              "FREE",
              "FREEMIUM",
              "LIMITED"
            ]
          },
          "active": {
            "type": "boolean"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "Payment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "invoice_id": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "COMPLETE",
              "INCOMPLETE"
            ]
          }
        }
      },
      "Photo":{
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          },
          "album_id": {
            "type": "string"
          },
          "captured_by": {
            "type": "string"
          },
          "captured_on": {
            "type": "string",
            "format": "date-time"
          },
          "captured_at": {
            "type": "string"
          },
          "resolution": {
            "type": "string"
          },
          "other_meta_data": {
            "type": "string"
          }
        }
      },
      "Invite":{
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "ADMIN",
              "SUPERADMIN",
              "EXECUTIVEBOARDMEMBER"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "JOINREQUEST",
              "VERIFICATION"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "SENT",
              "VERIFIED",
              "EXPIRED"
            ]
          }
        }
      }
    }
  }
  